Mimic linter's config file search algorithm
Since ansible-lint is executed from the root directory of the workspace
the algorithm that it uses to find configuration for a particular file
is not involved. To compensate, the extension now mimics that behavior,
searching the directory structure, going up from the investigated file.

At the same time, it is still possible to force a particular config file
through arguments provided in the extension settings.

The configuration file that is actually used is now correctly identified
for the purpose of marking configured finding groups as warnings.